Move zipkin constants to an internal package (#3431)
