+++ translator/trace/internal/zipkin/attributes.go
