+++ translator/trace/zipkinv1/json_test.go
@@ -31,7 +31,7 @@ import (
 
 	"go.opentelemetry.io/collector/translator/conventions"
 	tracetranslator "go.opentelemetry.io/collector/translator/trace"
-	"go.opentelemetry.io/collector/translator/trace/zipkin"
+	"go.opentelemetry.io/collector/translator/trace/internal/zipkin"
 )
 
 func TestSingleJSONV1BatchToTraces(t *testing.T) {
